backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv
alembic
httpx
python-multipart
pydantic>=2.0,<3.0
pydantic-settings
email-validator
fastapi-mail

#  Required for Google OAuth
google-auth
google-auth-oauthlib
google-auth-httplib2

#  Required for Replicate API
replicate
